AGENT-80: Sync metrics conversion updates
