+++ README.md
@@ -12,23 +12,23 @@ PremiScale uses [libvirt](https://libvirt.org/) to connect to hosts and manage l
 
 The controller can be configured to run in two different modes, including `kubernetes` (the default) and `standalone` modes. In either configuration, the controller aims to start only relevant processes for both data collection and managing virtual machines. Users are required at this time to provide a list of hosts on which virtual machines can be created, in addition to a list of autoscaling groups, into which virtual machines the controller manages, are organized.
 
+### Standalone
+
+In `standalone` mode (the default), the controller starts its own time series data collection process.
+
+<p align="center" width="100%">
+  <img width="100%" src="https://raw.githubusercontent.com/premiscale/premiscale/master/img/premiscale-architecture-controller_internal_autoscaler_enabled.png" alt="premiscale architecture: internal autoscaler enabled">
+</p>
+
 ### Kubernetes
 
-Starting the controller in `kubernetes` mode (the default) starts relevant components of the controller that allow it to interface with the cluster autoscaler.
+Starting the controller in `kubernetes` mode starts relevant components of the controller that allow it to interface with the cluster autoscaler.
 
 <p align="center" width="100%">
   <img width="100%" src="https://raw.githubusercontent.com/premiscale/premiscale/master/img/premiscale-architecture-controller_internal_autoscaler_disabled.png" alt="premiscale architecture: internal autoscaler disabled">
 </p>
 
 Note that, in this configuration, the controller does not require a time series database. State is still reconciled, but the time series signal comes from the cluster autoscaler instead.
-
-### Standalone
-
-In `standalone` mode, the controller starts its own time series data collection process.
-
-<p align="center" width="100%">
-  <img width="100%" src="https://raw.githubusercontent.com/premiscale/premiscale/master/img/premiscale-architecture-controller_internal_autoscaler_enabled.png" alt="premiscale architecture: internal autoscaler enabled">
-</p>
 
 ## Configuration
