+++ pyproject.toml
@@ -65,6 +65,7 @@ wrapt = "^1.16.0"
 xmltodict = "^0.13.0"
 tinyflux = "^1.0.0"
 grpcio = "^1.64.1"
+cachetools = "^5.3.3"
 
 [tool.poetry.group.profile.dependencies]
 memray = "^1.12"
