+++ requirements.txt
@@ -7,6 +7,9 @@ attrs==23.2.0 ; python_version >= "3.10" and python_version < "4.0" \
 blinker==1.8.2 ; python_version >= "3.10" and python_version < "4.0" \
     --hash=sha256:1779309f71bf239144b9399d06ae925637cf6634cf6bd131104184531bf67c01 \
     --hash=sha256:8f77b09d3bf7c795e969e9486f39c2c5e9c39d4ee07424be2bc594ece9642d83
+cachetools==5.3.3 ; python_version >= "3.10" and python_version < "4.0" \
+    --hash=sha256:0abad1021d3f8325b2fc1d2e9c8b9c9d57b04c3932657a72465447332c24d945 \
+    --hash=sha256:ba29e2dfa0b8b556606f097407ed1aa62080ee108ab0dc5ec9d6a723a007d105
 cattrs==23.2.3 ; python_version >= "3.10" and python_version < "4.0" \
     --hash=sha256:0341994d94971052e9ee70662542699a3162ea1e0c62f7ce1b4a57f563685108 \
     --hash=sha256:a934090d95abaa9e911dac357e3a8699e0b4b14f8529bcc7d2b1ad9d51672b9f
