+++ src/premiscale/hypervisor/_base.py
@@ -106,19 +106,6 @@ class Libvirt(ABC):
             log.error(f'No host connection to close, probably due to an error on connection open')
 
     @abstractmethod
-    def _getHostState(self) -> Dict:
-        """
-        Get the state of the VMs on the host.
-
-        Returns:
-            Dict: The state of the VMs on the host.
-
-        Raises:
-            NotImplementedError: If the method is not implemented by the subclass.
-        """
-        raise NotImplementedError
-
-    @abstractmethod
     def _getHostStats(self) -> Dict:
         """
         Get a report of schedulable resource utilization on the host.
@@ -145,19 +132,31 @@ class Libvirt(ABC):
         raise NotImplementedError
 
     @abstractmethod
-    def statsToStateDB(self) -> None:
+    def statsToStateDB(self) -> Dict:
         """
         Convert the stats from the host into a state database entry. Instead of relying on the calling class to
         format these correctly, every interface is required to implement its own method to do so, since it's not
         guaranteed that the stats will be the same across different hypervisors.
+
+        Returns:
+            Dict: The state of the host and VMs on it.
+
+        Raises:
+            NotImplementedError: If the method is not implemented by the subclass.
         """
-        return None
+        raise NotImplementedError
 
     @abstractmethod
-    def statsToMetricsDB(self) -> None:
+    def statsToMetricsDB(self) -> Dict:
         """
         Convert the stats from the host into a metrics database entry. Instead of relying on the calling class to
         format these correctly, every interface is required to implement its own method to do so, since it's not
         guaranteed that the stats will be the same across different hypervisors.
+
+        Returns:
+            Dict: The metrics for the host and VMs on it.
+
+        Raises:
+            NotImplementedError: If the method is not implemented by the subclass.
         """
-        return None
+        raise NotImplementedError
